fixed bar lines, cleaned up the bar line creation code a bit

+++ src/gameplay/beatmap_structs/timing_point.rs
@@ -1,10 +1,12 @@
 
 use std::sync::Arc;
 
+//TODO! implement time signature, its used for barlines
+
 #[derive(Clone)]
 pub struct TimingPoint {
     /// Start time of the timing section, in milliseconds from the beginning of the beatmap's audio. The end of the timing section is the next timing point's time (or never, if this is the last timing point).
-    pub time: u64,
+    pub time: f64,
     /// This property has two meanings:
     ///     For uninherited timing points, the duration of a beat, in milliseconds.
     ///     For inherited timing points, a negative inverse slider velocity multiplier, as a percentage. For example, -50 would make all sliders in this timing section twice as fast as SliderMultiplier.
@@ -19,7 +21,7 @@ impl TimingPoint {
         // time,beatLength,meter,sampleSet,sampleIndex,volume,uninherited,effects
         // println!("{}", str.clone());
         let mut split = str.split(',');
-        let time = split.next().unwrap().parse::<f64>().unwrap().round() as u64;
+        let time = split.next().unwrap().parse::<f64>().unwrap().round();
         let beat_length = split.next().unwrap().parse::<f32>().unwrap();
         let _meter = split.next().unwrap().parse::<u32>().unwrap();
         let _sample_set = split.next().unwrap().parse::<u32>().unwrap();
